Moved scripts to the scripts folder, made a new scene
Gonna be working on UI now. Will branch to new branch.
