+++ Assets/Eric's testing Area/MainUIScreen.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 239e824e0d0aadb4fb156c0326cc5853
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
