+++ Assets/Scripts/DataStructures.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 531f5edc5a86bfd4fbb470d1016f95ad
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
