+++ Assets/Scripts/DataStructures/EventNode.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 71ec344c0a875db49a0a4caf6325631c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
