+++ UserSettings/EditorUserSettings.asset
@@ -11,6 +11,9 @@ EditorUserSettings:
     RecentlyUsedScenePath-1:
       value: 224247031146467e1f070f6503634b15050311242b68072228285236f0e93707f6e225fdd8293c393707a1290a371526
       flags: 0
+    RecentlyUsedScenePath-2:
+      value: 224247031146467e1f070f6503634b15050311242b6807222828521ee3e93a0dcbd435fbe23f31722c0ce6281d
+      flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
       flags: 0
